fix: some compiler warnings
