+++ crates/bitwarden-iter/Cargo.toml
@@ -16,7 +16,9 @@ development = ["tokio-test"] # only used in doc-tests
 wasm = ["dep:wasm-bindgen", "dep:wasm-bindgen-futures"]
 
 [dependencies]
+futures = "0.3.31"
 wasm-bindgen = { workspace = true, optional = true }
+wasm-bindgen-futures = { workspace = true, optional = true }
 
 [target.'cfg(target_arch="wasm32")'.dependencies]
 wasm-bindgen-futures = { workspace = true, optional = true }
